data_processor/requirements.txt
fastapi
uvicorn
numpy
pillow
pandas